fix: add hack to fix inferred types
